// src/templates/template.enum.ts
export enum TemplateCategory {
  BUSINESS = 'business',
  EMPLOYMENT = 'employment',
  REAL_ESTATE = 'real_estate',
  INTELLECTUAL_PROPERTY = 'intellectual_property',
  FINANCIAL = 'financial',
  CUSTOM = 'custom',
  HEALTHCARE = 'healthcare',
  TECHNOLOGY = 'technology',
  ACADEMIC = 'academic',
  MEDIA = 'media',
}